+++ packages/zosmf/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to the Zowe z/OSMF SDK package will be documented in this file.
 
+## Recent Changes
+
+- Enhancement: Added `ZosmfChangePassword.changePassword` function which allows changing the z/OSMF password or passphrase. [#2703](https://github.com/zowe/zowe-cli/pull/2703)
+
 ## `8.27.4`
 
 - BugFix: Updated minimum supported version of Node from 18 to 20. Added Node 24 support. [#2616](https://github.com/zowe/zowe-cli/pull/2616)